+++ cognite_toolkit/_cdf_tk/resource_classes/instance.py
@@ -58,3 +58,36 @@ class NodeYAML(ToolkitResource):
         description="Fail the ingestion request if the node version is greater than this value.",
         ge=0,
     )
+
+
+class EdgeYAML(ToolkitResource):
+    space: str = Field(
+        description="Id of the space that the edge belongs to.",
+        min_length=1,
+        max_length=43,
+        pattern=SPACE_FORMAT_PATTERN,
+    )
+    external_id: str = Field(
+        description="External-id of the edge.",
+        min_length=1,
+        max_length=256,
+        pattern=r"^[^\\x00]{1,256}$",
+    )
+    type: InstanceType = Field(
+        description="The type of the edge, defined by a direct relation to another edge.",
+    )
+    sources: list[InstanceSource] | None = Field(
+        default=None,
+        description="List of sources (views) and their properties for this edge.",
+    )
+    existing_version: int | None = Field(
+        default=None,
+        description="Version of the edge.",
+        ge=0,
+    )
+    start_node: InstanceType = Field(
+        description="Reference to the node at the start of the edge.",
+    )
+    end_node: InstanceType = Field(
+        description="Reference to the node at the end of the edge.",
+    )
